integration plugin init
