+++ lib/plugins/_plugins.mjs
@@ -15,6 +15,7 @@ import { userLocale } from './userLocale.mjs';
 import { zoomBtn } from './zoomBtn.mjs';
 import { zoomToArea } from './zoomToArea.mjs';
 import { link_button } from './link_button.mjs';
+import { test } from './test.mjs';
 /**
 @typedef {Object} plugins
 @property {function} admin The admin plugin module.
@@ -41,6 +42,7 @@ const plugins = {
   zoomBtn,
   zoomToArea,
   link_button,
+  test,
 };
 
 export default plugins;
