+++ package.json
@@ -40,7 +40,7 @@
     "@eslint/js": "^9.13.0",
     "browser-sync": "^3.0.3",
     "clean-jsdoc-theme": "^4.3.0",
-    "codi-test-framework": "^1.0.24",
+    "codi-test-framework": "^1.0.34",
     "concurrently": "^9.1.0",
     "cookie-parser": "^1.4.5",
     "dotenv": "^16.4.5",
