+++ tests/browser/integration/_integration.test.mjs
@@ -1,0 +1,7 @@
+import { layerTest } from './layer.test.mjs';
+import { workspaceTest } from './workspace.test.mjs';
+
+export const integrationTests = {
+  layerTest,
+  workspaceTest,
+};
